fix(package): renamed package name
We have some libraries which are used this package name, so to make this work with them we need to
use another package name
