+++ android/src/main/AndroidManifest.xml
@@ -1,6 +1,5 @@
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.shahenlibrary">
+          package="com.extendable">
 
 </manifest>
-
