+++ android/src/main/java/com/extendable/ExtendableImage.java
@@ -1,4 +1,4 @@
-package com.shahenlibrary;
+package com.extendable;
 
 import android.content.Context;
 import android.util.AttributeSet;
