+++ android/src/main/java/com/extendable/ExtendableImageManager.java
@@ -1,5 +1,5 @@
 
-package com.shahenlibrary;
+package com.extendable;
 
 import android.support.annotation.Nullable;
 import android.util.Log;
