+++ android/src/main/java/com/extendable/RNExtendableImagePackage.java
@@ -1,5 +1,5 @@
 
-package com.shahenlibrary;
+package com.extendable;
 
 import java.util.Arrays;
 import java.util.Collections;
