+++ android/src/main/java/com/extendable/SendEventTypes.java
@@ -1,4 +1,4 @@
-package com.shahenlibrary;
+package com.extendable;
 
 /**
  * Created by Shahen on 6/30/17.
